pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-discord"
version = "0.2.0"
description = "Discord bot with Ollama integration supporting multiple configurations"
authors = [
    {name = "User", email = "user@example.com"}
]
readme = "CLAUDE.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0"
]

[project.scripts]
ollama-discord = "main:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.bandit]
exclude_dirs = ["tests"]